backend/requirements.txt
# requirements.txt

# Use the latest available versions
semantic-kernel==1.35.0
azure-ai-agents==1.1.0  
azure-identity==1.13.0
azure-ai-projects==1.1.0b1  # Use beta version

# fastapi
fastapi==0.104.1
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
numpy<2

# Azure Databricks with Unity Catalog support
databricks-sql-connector==3.0.0
azure-storage-blob==12.19.0  # For any blob storage operations if needed